Json is removed for resources

+++ polyescape_engine/src/polytech/teamf/api/ServiceServices.java
@@ -68,7 +68,7 @@ public class ServiceServices {
      */
     public String CaesarCipher(Object[] args) {
         WebTarget target = client.target(CAESAR_CIPHER_SERVICE_URI);
-        target = target.queryParam("message", (String) args[0]).queryParam("padding", (int) args[1]);
+        target = target.queryParam("message", (String) args[0]).queryParam("padding", args[1]);
         Invocation.Builder builder = target.request();
         builder.accept(MediaType.APPLICATION_JSON_TYPE);
         return builder.get(String.class);
